Concurrent Survey Process
Adds funcionality for CSP

https://trello.com/c/zgCyUiHl/204-3-concurrent-survey-process-step-3-ios-sdk

This commit:

- Includes end_user_id, user_id and unique_link on requests to responses and
decline
- Creates unique_link algorithm to be a SHA256 hash of NPS token,
end_user_email, date and a random string
- Adds priority number to responses and declines
- Adds WTRApiClientTests to test CSP changes

Code needed a refactor to make it more testable. I've focused on testing
the parts that involved CSP changes. There is still more testing to be
done but I prefered to leave that for another PR, atomic commits.

One thing I did include in this commit was changing identation of
ViewController.m and WTRApiClient.m from 4 spaces to 2. This is
and inconsistency throughout the project. I decided to settle for
2 spaces but haven't changed other files that use 4. Again, I'm
planning on doing this in a separate PR.

I've update the function to generate the unique link to use
a unix timestamp like date.

I've removed user_id test cases since it's no longer used.