Response/decline throttle
Updates the SDK to accept resurvey_throttle and decline_resurvey_throttle.
In case the end_user responds to the survey, the value of resurvey_throttle
will be used to set the expiration date of the surveyedAt value of
NSUserDefaults (cookie), resurvey_throttle default value is 90.
In case the end_user declines the survey, the value of decline_resurvey_throttle
will be used as the expiration date, default value is 30.

Other fixs:
-Fix last button long text adjustment
-Update tests
-Add tests for WTRDefaults
-DRY code

Trello: https://trello.com/c/hAeqz6Zr/733-1-custom-throttle-for-decline-response-ios-update-the-clients-need-the-new-response-resurvey-date-and-decline-resurvey-date-in-o