Fix bugs
This PR fixes the following bugs:
-SDK version bug
-Escape customProduct and customAudience
-Don't show survey if accountToken is wrong

*How to test*

1.-SDK version should be 0.5.9. Creating a new project
and adding WootricSDK should show this version and
not the project's version. sdk_version can be seen
on the request sent to eligibility on the logs.

2.-Setting a customProduct or customAudience with a
whitespace used to result on the request sent to
eligibility failing.

3.-accountToken being incorrect should show a white survey.
Testing can be done by using a correct clientId and
clientSecret. Also using  is good for testing.

Trello: https://trello.com/c/e4Ltv2Gr/812-1-5-ios-sdk-maintenance